+++ src/by/it/snegurskiy/jd02_05/Runner0205.java
@@ -19,20 +19,23 @@ public class Runner0205 {
 
     public static void main(String[] args) {
         if (args.length>1) Properties.setLocale(new Locale(args[0],args[1]));
-        Locale locale=new Locale("ru", "RU");
-        Properties.setLocale(locale);
-        System.out.println(Properties.getString("Hello"));
-        System.out.println(Properties.getString("Howdoyoudo")+"\n");
-        Properties.setLocale(new Locale("en","US"));
-        System.out.println(Properties.getString("Hello"));
-        System.out.println(Properties.getString("Howdoyoudo")+"\n");
-        Properties.setLocale(new Locale("be","BY"));
-        System.out.println(Properties.getString("Hello"));
-        System.out.println(Properties.getString("Howdoyoudo")+"\n");
+        Locale locale=new Locale("", "");
+//        Locale locale=new Locale("ru", "RU");
+//        Properties.setLocale(locale);
+//        System.out.println(Properties.getString("Hello"));
+//        System.out.println(Properties.getString("Howdoyoudo")+"\n");
+//        Properties.setLocale(new Locale("en","US"));
+//        System.out.println(Properties.getString("Hello"));
+//        System.out.println(Properties.getString("Howdoyoudo")+"\n");
+//        Properties.setLocale(new Locale("be","BY"));
+//        System.out.println(Properties.getString("Hello"));
+//        System.out.println(Properties.getString("Howdoyoudo")+"\n");
 
         DateFormat df=DateFormat.getDateInstance(DateFormat.LONG,Properties.getLocale());
         Date d=new Date(System.currentTimeMillis());
         System.out.println(df.format(d));
+
+
         System.out.println(Properties.getString("set.Language"));
         String language;
         while (!(language=getOneConsole()).isEmpty()) {
