+++ src/by/it/snegurskiy/jd02_05/Txt_be_BY.properties
@@ -1,3 +1,3 @@
-Hello = Прывітанне, Мікола!
-Howdoyoudo= як справы?
-set.Language=мова беларуская
+Hello = \u041f\u0440\u044b\u0432\u0456\u0442\u0430\u043d\u043d\u0435, \u041c\u0456\u043a\u043e\u043b\u0430!
+Howdoyoudo= \u044f\u043a \u0441\u043f\u0440\u0430\u0432\u044b?
+set.Language=\u043c\u043e\u0432\u0430 \u0431\u0435\u043b\u0430\u0440\u0443\u0441\u043a\u0430\u044f
