+++ src/by/it/snegurskiy/jd02_05/Txt_be_BY.txt
@@ -1,3 +1,3 @@
 Hello = Прывітанне, Мікола!
-Howdoeoudo= як справы?
+Howdoyoudo= як справы?
 set.Language=мова беларуская
