+++ src/by/it/snegurskiy/jd02_05/Txt_ru_RU.properties
@@ -1,3 +1,3 @@
-Hello = Привет, Николай!
-Howdoyoudo = Как дела?
-set.Language=Язык ввода русский
+Hello = \u041f\u0440\u0438\u0432\u0435\u0442, \u041d\u0438\u043a\u043e\u043b\u0430\u0439!
+Howdoyoudo = \u041a\u0430\u043a \u0434\u0435\u043b\u0430?
+set.Language=\u042f\u0437\u044b\u043a \u0432\u0432\u043e\u0434\u0430 \u0440\u0443\u0441\u0441\u043a\u0438\u0439
