+++ src/by/it/snegurskiy/jd03/classwork/CN.java
@@ -1,0 +1,7 @@
+package by.it.snegurskiy.jd03.classwork;
+
+/**
+ * Created by Acer on 21.10.2016.
+ */
+public class CN {
+}
