+++ src/by/it/snegurskiy/jd03/classwork/CRUD/UsersCRUD.java
@@ -1,4 +1,7 @@
-package by.it.snegurskiy.jd03_01.classwork.sql.classwork;
+package by.it.snegurskiy.jd03.classwork.CRUD;
+
+import by.it.snegurskiy.jd03.classwork.ConnectionCreator;
+import by.it.snegurskiy.jd03.classwork.Users;
 
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -6,84 +9,88 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Locale;
 
-public class PaymentsCRUD {
-    public Payments create(Payments payment) throws SQLException {
-        payment.setID(0);
+/**
+ * Created by Acer on 19.10.2016.
+ */
+public class UsersCRUD {
+
+    public Users create(Users user) throws SQLException {
+        user.setID(0);
         //формирование строки createUserSQL по данным bean user
-        String createPaymentsSQL = String.format(Locale.ENGLISH,
-                "insert into payment(Data, Type, Sum, Source, FK_account) values(%d,'%s',%d,%d,%d);",
-                payment.getData(),payment.getType(),payment.getSum(), payment.getSource(), payment.getFK_account()
+        String createUserSQL = String.format(Locale.ENGLISH,
+                "insert into users(Surname, Name, Login,Password,Email,FK_Role) values('%s','%s','%s','%s','%s',%d);",
+                user.getSurname(),user.getName(),user.getLogin(), user.getPassword(), user.getFK_Role()
         );
         try (
                 //создаем соединение и объект для запросов к базе
-                Connection connection=ConnectionCreator.getConnection();
+                Connection connection= ConnectionCreator.getConnection();
 //                Connection connection = ConnectionCreator.getConnection(); //создаем соединение с базой
                 Statement statement = connection.createStatement(); //создаем объект для обращения к базе
         ) {
             //выполняем добавление в базу, должна быть добавлена одна запись. Проверим это.
             //create(insert) update delete - это executeUpdate, а select это executeQuery
-            if (statement.executeUpdate(createPaymentsSQL) == 1)
+            if (statement.executeUpdate(createUserSQL) == 1)
             {
                 //если все добавлено то узнаем последний ID
                 ResultSet resultSet = statement.executeQuery("SELECT LAST_INSERT_ID();");
                 //извлекаем из resultSet последний ID
                 if (resultSet.next())
-                    payment.setID(resultSet.getInt(1));
+                    user.setID(resultSet.getInt(1));
             }
         } catch (SQLException e) {
             throw e;
         }
-        return payment;
+        return user;
     }
 
-    public Payments read(int id) throws SQLException {
-        Payments paymentResult = null;
-        String readPaymentsSQL = "SELECT * FROM payments where ID=" + id;
+    public Users read(int id) throws SQLException {
+        Users userResult = null;
+        String readUserSQL = "SELECT * FROM users where ID=" + id;
         try (
                 Connection connection = ConnectionCreator.getConnection();
                 Statement statement = connection.createStatement();
         ) {
-            final ResultSet resultSet = statement.executeQuery(readPaymentsSQL);
+            final ResultSet resultSet = statement.executeQuery(readUserSQL);
             if (resultSet.next()) {
-                paymentResult = new Payments(
+                userResult = new Users(
                         resultSet.getInt("ID"),
-                        resultSet.getInt("Data"),
-                        resultSet.getString("Type"),
-                        resultSet.getDouble("Sum"),
-                        resultSet.getInt("Source"),
-                        resultSet.getInt("FK_account"));
+                        resultSet.getString("Surname"),
+                        resultSet.getString("Name"),
+                        resultSet.getString("Login"),
+                        resultSet.getString("Password"),
+                        resultSet.getInt("FK_Role"));
             }
         } catch (SQLException e) {
             throw e;
         }
-        return paymentResult;
+        return userResult;
     }
 
-    public Payments update(Payments payment) throws SQLException {
-        Payments paymentResult = null;
-        String updatePaymentsSQL = String.format(
-                "UPDATE payments SET Data = %d, Type = '%s',Sum = %d, Source = %d, FK_account=%d WHERE payments.ID = %d",
-                payment.getData(),payment.getType(),payment.getSum(), payment.getSource(), payment.getFK_account(), payment.getID()
+    public Users update(Users user) throws SQLException {
+        Users userResult = null;
+        String updateUserSQL = String.format(
+                "UPDATE users SET Surname = '%s',Name = '%s',Login = '%s', Password = '%s', FK_Role=%d WHERE users.ID = %d",
+                user.getSurname(),user.getName(),user.getLogin(), user.getPassword(), user.getFK_Role(), user.getID()
         );
         try (
                 Connection connection = ConnectionCreator.getConnection();
                 Statement statement = connection.createStatement();
         ) {
-            if (statement.executeUpdate(updatePaymentsSQL) == 1)
-                paymentResult = payment;
+            if (statement.executeUpdate(updateUserSQL) == 1)
+                userResult = user;
         } catch (SQLException e) {
             throw e;
         }
-        return paymentResult;
+        return userResult;
     }
 
-    public boolean delete(Payments payment) throws SQLException {
-        String deletePaymentsSQL = String.format("DELETE FROM payments WHERE payments.ID = %d", payment.getID());
+    public boolean delete(Users user) throws SQLException {
+        String deleteUserSQL = String.format("DELETE FROM users WHERE users.ID = %d", user.getID());
         try (
                 Connection connection = ConnectionCreator.getConnection();
                 Statement statement = connection.createStatement();
         ) {
-            return (statement.executeUpdate(deletePaymentsSQL) == 1);
+            return (statement.executeUpdate(deleteUserSQL) == 1);
         } catch (SQLException e) {
             throw e;
         }
