+++ src/by/it/snegurskiy/jd03/classwork/ConnectionCreator.java
@@ -1,4 +1,4 @@
-package by.it.snegurskiy.jd03_01.classwork.sql.classwork;
+package by.it.snegurskiy.jd03.classwork;
 
 import com.mysql.fabric.jdbc.FabricMySQLDriver;
