+++ src/by/it/snegurskiy/jd03/classwork/Payments.java
@@ -1,4 +1,4 @@
-package by.it.snegurskiy.jd03_01.classwork.sql.classwork;
+package by.it.snegurskiy.jd03.classwork;
 
 /**
  * Created by Acer on 21.10.2016.
