+++ src/by/it/snegurskiy/jd03/classwork/UsersTest.java
@@ -1,7 +1,8 @@
-package by.it.snegurskiy.jd03_01.classwork.sql.classwork;
+package by.it.snegurskiy.jd03.classwork;
 
 /**
  * Created by Acer on 19.10.2016.
  */
 public class UsersTest {
+
 }
